# SocialNetwork/src/main/java/kattsyn/dev/SocialNetwork/entities/Role.java
package kattsyn.dev.SocialNetwork.entities;


import jakarta.persistence.*;
import lombok.Data;

@Entity
@Data
@Table(name = "roles")
public class Role {
    @Id
    @GeneratedValue(strategy = GenerationType.IDENTITY)
    @Column(name = "id")
    private Integer id;

    @Column(name = "name")
    private String name;
}